--- src/main/kotlin/io/github/trcdevelopers/clayium/integration/gregtech/GTOreDictUnifierAdapter.kt
@@ -1,28 +1,0 @@
-package io.github.trcdevelopers.clayium.integration.gregtech
-
-import io.github.trcdevelopers.clayium.api.util.copyWithSize
-import io.github.trcdevelopers.clayium.common.unification.IOreDictUnifier
-import net.minecraft.item.ItemStack
-
-object GTOreDictUnifierAdapter : IOreDictUnifier {
-    override fun registerOre(stack: ItemStack, oreDict: String) {
-        gregtech.api.unification.OreDictUnifier.registerOre(stack, oreDict)
-    }
-
-    override fun getOreNames(stack: ItemStack): Set<String> {
-        return gregtech.api.unification.OreDictUnifier.getOreDictionaryNames(stack)
-    }
-
-    override fun get(oreDict: String, amount: Int): ItemStack {
-        return gregtech.api.unification.OreDictUnifier.get(oreDict)
-            .copyWithSize(amount)
-    }
-
-    override fun getAll(oreDict: String, amount: Int): List<ItemStack> {
-        return gregtech.api.unification.OreDictUnifier.getAllWithOreDictionaryName(oreDict)
-    }
-
-    override fun has(stack: ItemStack, oreDict: String): Boolean {
-        return gregtech.api.unification.OreDictUnifier.hasOreDictionary(stack, oreDict)
-    }
-}
